Add reusable checklist preview for inspector workflows
